arreglado popular product falta que muestre el json en la consulta

+++ shop-microservice/src/main/java/com/integrador5/shopmicroservice/DTO/MostPopularProductDTO.java
@@ -3,9 +3,9 @@ package com.integrador5.shopmicroservice.DTO;
 public class MostPopularProductDTO {
     private int idProduct;
     private String nameProduct;
-    private int quantity;
+    private long quantity;
 
-    public MostPopularProductDTO(int idProduct, String nameProduct, int quantity) {
+    public MostPopularProductDTO(int idProduct, String nameProduct, long quantity) {
         this.idProduct = idProduct;
         this.nameProduct = nameProduct;
         this.quantity = quantity;
